Refactor file structure and aquire_utils functions
