+++ tests/application_tests/measurement_tests/comparison_json/mixed_d2233_p20.json
