+++ tests/application_tests/measurement_tests/comparison_json/mixed_d235_p20.json
