+++ tests/application_tests/measurement_tests/comparison_json/mono_qubit_d2222_p40.json
