+++ tests/application_tests/measurement_tests/comparison_json/mono_qubit_d222_p10.json
