+++ tests/application_tests/measurement_tests/comparison_json/mono_qudit_d333_p20.json
